refactored a lot of the duplicated response classes
