+++ src/NswLicensePlateLookup/Services/PlateLookupService.cs
@@ -24,17 +24,18 @@ namespace NswLicensePlateLookup.Services
             var serviceNswRequestBody = new ServiceNswRequestBody();
             serviceNswRequestBody.Method = "createRMSTransaction";
             serviceNswRequestBody.Data = new List<string>{"{\"ipAddress\":\"164d9d47a20e4fe26aed22c7cd88e832ee012389c233fc756b3c98f53f5381e6\",\"transactionName\":\"FREEREGCHK\",\"outletNumber\":\"\"}"};
-            var tokenResponseList = await serviceApi.SendServiceNswRequest(serviceNswRequestBody);
+            var tokenResponseList = await serviceApi.SendServiceNswRequest<TokenResult>(serviceNswRequestBody);
             var tokenResponse = tokenResponseList[0];
-            var token = tokenResponse.Result.StatusObject;
+            var token = tokenResponse.Result.Token;
 
             // use token to request plate details
+            serviceNswRequestBody = new ServiceNswRequestBody();
             serviceNswRequestBody.Method = "postVehicleListForFreeRegoCheck";
             serviceNswRequestBody.Data = new List<string>{"{\"transactionToken\":\"" + token + "\",\"plateNumber\":\"" + plateNumber + "\"}"};
-            var plateDetailsResponseList = await serviceApi.SendServiceNswRequest2(serviceNswRequestBody);
+            var plateDetailsResponseList = await serviceApi.SendServiceNswRequest<PlateDetailsResult>(serviceNswRequestBody);
             var plateDetailsResponse = plateDetailsResponseList[0];
             
-            return plateDetailsResponse.Result.StatusObject.Vehicle.Model;
+            return plateDetailsResponse.Result.PlateDetails.Vehicle.Model;
         }
     }
 
@@ -42,11 +43,6 @@ namespace NswLicensePlateLookup.Services
     {
         [Headers("Content-Type: application/json", "origin: https://my.service.nsw.gov.au", "referer: https://my.service.nsw.gov.au/MyServiceNSW/index")]
         [Post("/MyServiceNSW/apexremote")]
-        Task<List<TokenResponse>> SendServiceNswRequest(ServiceNswRequestBody body);
-        
-        //lots of duplicated code here!!
-        [Headers("Content-Type: application/json", "origin: https://my.service.nsw.gov.au", "referer: https://my.service.nsw.gov.au/MyServiceNSW/index")]
-        [Post("/MyServiceNSW/apexremote")]
-        Task<List<PlateDetailsResponse>> SendServiceNswRequest2(ServiceNswRequestBody body);
+        Task<List<ServiceNswResponse<ResultType>>> SendServiceNswRequest<ResultType>(ServiceNswRequestBody body);
     }
 }
